resultディレクトリを自動生成する，README加筆

+++ execution.sh
@@ -1,2 +1,7 @@
-rm ./result/json/*
+if [ -e "./result" ]; then
+  rm ./result/json/*
+else
+  mkdir -p result/json result/img
+fi
+
 time ruby checkins.rb
